move GenesisState generators to modules
