+++ x/genaccounts/simulation/genesis.go
@@ -1,0 +1,75 @@
+package simulation
+
+// DONTCOVER
+
+import (
+	"encoding/json"
+	"math/rand"
+	"time"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/genaccounts"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+)
+
+// GenGenesisAccounts generates a random GenesisState for the genesis accounts
+func GenGenesisAccounts(
+	cdc *codec.Codec, r *rand.Rand, accs []simulation.Account,
+	genesisTimestamp time.Time, amount, numInitiallyBonded int64,
+	genesisState map[string]json.RawMessage,
+) {
+
+	var genesisAccounts []genaccounts.GenesisAccount
+
+	// randomly generate some genesis accounts
+	for i, acc := range accs {
+		coins := sdk.Coins{sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(amount))}
+		bacc := auth.NewBaseAccountWithAddress(acc.Address)
+		bacc.SetCoins(coins)
+
+		var gacc genaccounts.GenesisAccount
+
+		// Only consider making a vesting account once the initial bonded validator
+		// set is exhausted due to needing to track DelegatedVesting.
+		if int64(i) > numInitiallyBonded && r.Intn(100) < 50 {
+			var (
+				vacc    auth.VestingAccount
+				endTime int64
+			)
+
+			startTime := genesisTimestamp.Unix()
+
+			// Allow for some vesting accounts to vest very quickly while others very slowly.
+			if r.Intn(100) < 50 {
+				endTime = int64(simulation.RandIntBetween(r, int(startTime), int(startTime+(60*60*24*30))))
+			} else {
+				endTime = int64(simulation.RandIntBetween(r, int(startTime), int(startTime+(60*60*12))))
+			}
+
+			if startTime == endTime {
+				endTime++
+			}
+
+			if r.Intn(100) < 50 {
+				vacc = auth.NewContinuousVestingAccount(&bacc, startTime, endTime)
+			} else {
+				vacc = auth.NewDelayedVestingAccount(&bacc, endTime)
+			}
+
+			var err error
+			gacc, err = genaccounts.NewGenesisAccountI(vacc)
+			if err != nil {
+				panic(err)
+			}
+		} else {
+			gacc = genaccounts.NewGenesisAccount(&bacc)
+		}
+
+		genesisAccounts = append(genesisAccounts, gacc)
+	}
+
+	genesisState[genaccounts.ModuleName] = cdc.MustMarshalJSON(genesisAccounts)
+}
