+++ x/slashing/simulation/genesis.go
@@ -1,0 +1,74 @@
+package simulation
+
+// DONTCOVER
+
+import (
+	"encoding/json"
+	"fmt"
+	"math/rand"
+	"time"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+	"github.com/cosmos/cosmos-sdk/x/slashing"
+	"github.com/cosmos/cosmos-sdk/x/staking"
+)
+
+// GenSlashingGenesisState generates a random GenesisState for slashing
+func GenSlashingGenesisState(
+	cdc *codec.Codec, r *rand.Rand, stakingGen staking.GenesisState,
+	ap simulation.AppParams, genesisState map[string]json.RawMessage,
+) {
+	slashingGenesis := slashing.NewGenesisState(
+		slashing.NewParams(
+			stakingGen.Params.UnbondingTime,
+			func(r *rand.Rand) int64 {
+				var v int64
+				ap.GetOrGenerate(cdc, simulation.SignedBlocksWindow, &v, r,
+					func(r *rand.Rand) {
+						v = simulation.ModuleParamSimulator[simulation.SignedBlocksWindow](r).(int64)
+					})
+				return v
+			}(r),
+			func(r *rand.Rand) sdk.Dec {
+				var v sdk.Dec
+				ap.GetOrGenerate(cdc, simulation.MinSignedPerWindow, &v, r,
+					func(r *rand.Rand) {
+						v = simulation.ModuleParamSimulator[simulation.MinSignedPerWindow](r).(sdk.Dec)
+					})
+				return v
+			}(r),
+			func(r *rand.Rand) time.Duration {
+				var v time.Duration
+				ap.GetOrGenerate(cdc, simulation.DowntimeJailDuration, &v, r,
+					func(r *rand.Rand) {
+						v = simulation.ModuleParamSimulator[simulation.DowntimeJailDuration](r).(time.Duration)
+					})
+				return v
+			}(r),
+			func(r *rand.Rand) sdk.Dec {
+				var v sdk.Dec
+				ap.GetOrGenerate(cdc, simulation.SlashFractionDoubleSign, &v, r,
+					func(r *rand.Rand) {
+						v = simulation.ModuleParamSimulator[simulation.SlashFractionDoubleSign](r).(sdk.Dec)
+					})
+				return v
+			}(r),
+			func(r *rand.Rand) sdk.Dec {
+				var v sdk.Dec
+				ap.GetOrGenerate(cdc, simulation.SlashFractionDowntime, &v, r,
+					func(r *rand.Rand) {
+						v = simulation.ModuleParamSimulator[simulation.SlashFractionDowntime](r).(sdk.Dec)
+					})
+				return v
+			}(r),
+		),
+		nil,
+		nil,
+	)
+
+	fmt.Printf("Selected randomly generated slashing parameters:\n%s\n", codec.MustMarshalJSONIndent(cdc, slashingGenesis.Params))
+	genesisState[slashing.ModuleName] = cdc.MustMarshalJSON(slashingGenesis)
+}
